Refactor schedules into service module
