+++ tracecat/workflow/schedules/router.py
@@ -5,12 +5,11 @@ from sqlalchemy.exc import NoResultFound
 from sqlmodel import select
 from sqlmodel.ext.asyncio.session import AsyncSession
 
-from tracecat import identifiers
 from tracecat.auth.credentials import TemporaryRole, authenticate_user_for_workspace
 from tracecat.db.engine import get_async_session
 from tracecat.db.schemas import Schedule, WorkflowDefinition
-from tracecat.dsl import schedules
 from tracecat.dsl.common import DSLInput
+from tracecat.identifiers import ScheduleID, WorkflowID
 from tracecat.logging import logger
 from tracecat.types.api import (
     CreateScheduleParams,
@@ -18,15 +17,19 @@ from tracecat.types.api import (
     UpdateScheduleParams,
 )
 from tracecat.types.auth import Role
+from tracecat.workflow.schedules import service
 
 router = APIRouter(prefix="/schedules")
+
+WorkspaceUserRole = Annotated[Role, Depends(authenticate_user_for_workspace)]
+AsyncDBSession = Annotated[AsyncSession, Depends(get_async_session)]
 
 
 @router.get("", tags=["schedules"])
 async def list_schedules(
-    role: Annotated[Role, Depends(authenticate_user_for_workspace)],
-    workflow_id: identifiers.WorkflowID | None = None,
-    session: AsyncSession = Depends(get_async_session),
+    role: WorkspaceUserRole,
+    session: AsyncDBSession,
+    workflow_id: WorkflowID | None = None,
 ) -> list[Schedule]:
     """List all schedules for a workflow."""
     statement = select(Schedule).where(Schedule.owner_id == role.workspace_id)
@@ -39,77 +42,74 @@ async def list_schedules(
 
 @router.post("", tags=["schedules"])
 async def create_schedule(
-    role: Annotated[Role, Depends(authenticate_user_for_workspace)],
+    role: WorkspaceUserRole,
+    session: AsyncDBSession,
     params: CreateScheduleParams,
-    session: AsyncSession = Depends(get_async_session),
 ) -> Schedule:
     """Create a schedule for a workflow."""
 
-    with logger.contextualize(role=role):
-        result = await session.exec(
-            select(WorkflowDefinition)
-            .where(WorkflowDefinition.workflow_id == params.workflow_id)
-            .order_by(WorkflowDefinition.version.desc())
-        )
-        try:
-            if not (defn_data := result.first()):
-                raise NoResultFound("No workflow definition found for workflow ID")
-        except NoResultFound as e:
-            logger.opt(exception=e).error("Invalid workflow ID", error=e)
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND, detail="Invalid workflow ID"
-            ) from e
+    result = await session.exec(
+        select(WorkflowDefinition)
+        .where(WorkflowDefinition.workflow_id == params.workflow_id)
+        .order_by(WorkflowDefinition.version.desc())
+    )
+    try:
+        if not (defn_data := result.first()):
+            raise NoResultFound("No workflow definition found for workflow ID")
+    except NoResultFound as e:
+        logger.opt(exception=e).error("Invalid workflow ID", error=e)
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Invalid workflow ID"
+        ) from e
 
-        schedule = Schedule(
-            owner_id=role.workspace_id, **params.model_dump(exclude_unset=True)
-        )
-        await session.refresh(defn_data)
-        defn = WorkflowDefinition.model_validate(defn_data)
-        dsl = DSLInput(**defn.content)
+    schedule = Schedule(
+        owner_id=role.workspace_id, **params.model_dump(exclude_unset=True)
+    )
+    await session.refresh(defn_data)
+    defn = WorkflowDefinition.model_validate(defn_data)
+    dsl = DSLInput(**defn.content)
 
-        try:
-            # Set the role for the schedule as the tracecat-runner
-            with TemporaryRole(
-                type="service",
-                user_id=defn.owner_id,
-                service_id="tracecat-schedule-runner",
-            ) as sch_role:
-                handle = await schedules.create_schedule(
-                    workflow_id=params.workflow_id,
-                    schedule_id=schedule.id,
-                    dsl=dsl,
-                    every=params.every,
-                    offset=params.offset,
-                    start_at=params.start_at,
-                    end_at=params.end_at,
-                    trigger_inputs=params.inputs,
-                )
-                logger.info(
-                    "Created schedule",
-                    handle_id=handle.id,
-                    workflow_id=params.workflow_id,
-                    schedule_id=schedule.id,
-                    sch_role=sch_role,
-                )
+    try:
+        # Set the role for the schedule as the tracecat-runner
+        with TemporaryRole(
+            type="service",
+            user_id=defn.owner_id,
+            service_id="tracecat-schedule-runner",
+        ) as sch_role:
+            handle = await service.create_schedule(
+                workflow_id=params.workflow_id,
+                schedule_id=schedule.id,
+                dsl=dsl,
+                every=params.every,
+                offset=params.offset,
+                start_at=params.start_at,
+                end_at=params.end_at,
+                trigger_inputs=params.inputs,
+            )
+            logger.info(
+                "Created schedule",
+                handle_id=handle.id,
+                workflow_id=params.workflow_id,
+                schedule_id=schedule.id,
+                sch_role=sch_role,
+            )
 
-            session.add(schedule)
-            await session.commit()
-            await session.refresh(schedule)
-            return schedule
-        except Exception as e:
-            session.rollback()
-            logger.opt(exception=e).error("Error creating schedule", error=e)
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail="Error creating schedule",
-            ) from e
+        session.add(schedule)
+        await session.commit()
+        await session.refresh(schedule)
+        return schedule
+    except Exception as e:
+        session.rollback()
+        logger.opt(exception=e).error("Error creating schedule", error=e)
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail="Error creating schedule",
+        ) from e
 
 
 @router.get("/{schedule_id}", tags=["schedules"])
 async def get_schedule(
-    role: Annotated[Role, Depends(authenticate_user_for_workspace)],
-    schedule_id: identifiers.ScheduleID,
-    session: AsyncSession = Depends(get_async_session),
+    role: WorkspaceUserRole, session: AsyncDBSession, schedule_id: ScheduleID
 ) -> Schedule:
     """Get a schedule from a workflow."""
     statement = select(Schedule).where(
@@ -120,17 +120,17 @@ async def get_schedule(
         schedule = result.one()
     except NoResultFound as e:
         raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Resource not found"
+            status_code=status.HTTP_404_NOT_FOUND, detail="Schedule not found"
         ) from e
     return schedule
 
 
 @router.post("/{schedule_id}", tags=["schedules"])
 async def update_schedule(
-    role: Annotated[Role, Depends(authenticate_user_for_workspace)],
-    schedule_id: identifiers.ScheduleID,
+    role: WorkspaceUserRole,
+    session: AsyncDBSession,
+    schedule_id: ScheduleID,
     params: UpdateScheduleParams,
-    session: AsyncSession = Depends(get_async_session),
 ) -> Schedule:
     """Update a schedule from a workflow. You cannot update the Workflow Definition, but you can update other fields."""
     statement = select(Schedule).where(
@@ -141,12 +141,12 @@ async def update_schedule(
         schedule = result.one()
     except NoResultFound as e:
         raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Resource not found"
+            status_code=status.HTTP_404_NOT_FOUND, detail="Schedule not found"
         ) from e
 
     try:
         # (1) Synchronize with Temporal
-        await schedules.update_schedule(schedule_id, params)
+        await service.update_schedule(schedule_id, params)
 
         # (2) Update the schedule
         for key, value in params.model_dump(exclude_unset=True).items():
@@ -172,9 +172,7 @@ async def update_schedule(
     tags=["schedules"],
 )
 async def delete_schedule(
-    role: Annotated[Role, Depends(authenticate_user_for_workspace)],
-    schedule_id: identifiers.ScheduleID,
-    session: AsyncSession = Depends(get_async_session),
+    role: WorkspaceUserRole, session: AsyncDBSession, schedule_id: ScheduleID
 ) -> None:
     """Delete a schedule from a workflow."""
     statement = select(Schedule).where(
@@ -190,7 +188,7 @@ async def delete_schedule(
 
     try:
         # Delete the schedule from Temporal first
-        await schedules.delete_schedule(schedule_id)
+        await service.delete_schedule(schedule_id)
 
         # If successful, delete the schedule from the database
         if schedule:
@@ -211,9 +209,7 @@ async def delete_schedule(
 
 @router.get("/search", tags=["schedules"])
 async def search_schedules(
-    role: Annotated[Role, Depends(authenticate_user_for_workspace)],
-    params: SearchScheduleParams,
-    session: AsyncSession = Depends(get_async_session),
+    role: WorkspaceUserRole, session: AsyncDBSession, params: SearchScheduleParams
 ) -> list[Schedule]:
     """**[WORK IN PROGRESS]** Search for schedules."""
     statement = select(Schedule).where(Schedule.owner_id == role.workspace_id)
